mcp-server-openmetadata/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
mcp==1.2.1
-e /home/team/MCP-Server/mcp-server-openmetadata
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.45.3
typing-extensions==4.12.2
uvicorn==0.34.0
